improvement(tools/quantize): refactor code
