+++ tools/quantize/net_quantize.h
@@ -5,15 +5,14 @@
 class NetQuantize : public ModelWriter
 {
 public:
-    NetQuantize();
+    NetQuantize() {}
 
     std::map<std::string, ncnn::Mat> blob_int8scale_table;
     std::map<std::string, ncnn::Mat> weight_int8scale_table;
 
 public:
-    void visit(const cpptoml::value<std::string>& v);
-    void visit(const cpptoml::table& t);
-    void read_toml(const std::string& toml_path);
+    bool read_raw_format(const char* path);
+    bool read_ini_format(const char* path);
     int quantize_convolution();
     int quantize_convolutiondepthwise();
     int quantize_innerproduct();
